Update spinlock fairness test to use C++17
